fix: refactoring files and file names - no code or features updated
