+++ lib/screens/login.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:runfinity/styles/app_colors.dart';
-import 'package:runfinity/widgets/appText.dart';
-import 'package:runfinity/utils/formValidation.dart';
-import 'package:runfinity/widgets/login/loginCheckBox.dart';
-import 'package:runfinity/widgets/login/loginInput.dart';
-import 'package:runfinity/widgets/login/loginPasswordInput.dart';
-import 'package:runfinity/widgets/login/otherLoginOptions.dart';
+import 'package:runfinity/widgets/app_text.dart';
+import 'package:runfinity/utils/form_validation.dart';
+import 'package:runfinity/widgets/check_box.dart';
+import 'package:runfinity/widgets/login/login_input.dart';
+import 'package:runfinity/widgets/login/login_password_input.dart';
+import 'package:runfinity/widgets/login/other_login_options.dart';
 
 class Login extends StatefulWidget {
   const Login({super.key});
@@ -73,7 +73,7 @@ class _LoginState extends State<Login> {
                               children: [
                                 Row(
                                   children: [
-                                    const LoginCheckbox(),
+                                    const CheckBox(),
                                     AppText(text: 'Remember Me')
                                   ],
                                 ),
