+++ lib/screens/signup.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:runfinity/styles/app_colors.dart';
-import 'package:runfinity/widgets/appText.dart';
-import 'package:runfinity/utils/formValidation.dart';
-import 'package:runfinity/widgets/login/loginInput.dart';
-import 'package:runfinity/widgets/login/loginPasswordInput.dart';
-import 'package:runfinity/widgets/login/otherLoginOptions.dart';
+import 'package:runfinity/widgets/app_text.dart';
+import 'package:runfinity/utils/form_validation.dart';
+import 'package:runfinity/widgets/login/login_input.dart';
+import 'package:runfinity/widgets/login/login_password_input.dart';
+import 'package:runfinity/widgets/login/other_login_options.dart';
 
 class SignUp extends StatefulWidget {
   const SignUp({super.key});
