+++ lib/widgets/lobby/create_lobby_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../styles/app_colors.dart';
+import '../../styles/app_colors.dart';
 
 class CreateLobbyScreen extends StatefulWidget {
   const CreateLobbyScreen({super.key});
