+++ lib/widgets/login/other_login_options.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:runfinity/styles/app_colors.dart';
-import 'package:runfinity/widgets/appText.dart';
+import 'package:runfinity/widgets/app_text.dart';
 
 class OtherLoginOptions extends StatelessWidget {
   const OtherLoginOptions({super.key});
