+++ lib/widgets/profile/profile_name_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:runfinity/styles/app_colors.dart';
-import 'package:runfinity/widgets/appText.dart';
+import 'package:runfinity/widgets/app_text.dart';
 
 class ProfileNameCard extends StatelessWidget {
   const ProfileNameCard({super.key});
